Fix system_get_status stat2 chip mode return
According to the official documentation of the LR1110 chip, the Stat2
register consists of the following:

------------------------------------------
|     [7:4]    |    [3:1]   |     [0]    |
------------------------------------------
| Reset status | Chip mode  | Bootloader |
------------------------------------------

However, when the chip_mode field in stat2 is assigned, the entire upper
value (0xFE), shifted one bit right is taken:

`stat2->chip_mode = (cbuffer[1] >> 1);`

Because of this we also read the 'Reset status' bits, giving us a wrong
'Chip mode' value if the device was reset beforehand.

In order to match the documentation, the code should first select the
chip_mode bits and then shift the value to the right:

`stat2->chip_mode = ((cbuffer[1] & 0x0F) >> 1);`

It might also be a good idea to create an additional field in
`lr1110_system_stat2_t` called `reset_status` and assign the upper 4
bits to it:

`stat2->reset_status = ((cbuffer[1] & 0xF0) >> 4);`

I will leave the above up to the makers of the drivers.

See: 'LR1110 User Manual Rev.1.0 March 2020' § 3.3.2.